(chore) extract handleFormBlur from state helpers file and move it to components/Login/helpers, rename state hook file to useLogState
