+++ frontend/src/components/login/helpers/handle-form-blur.ts
@@ -1,0 +1,7 @@
+/**
+ * Simple form input blur handler. Only works on top-level keys, don't use for nested form state.
+ */
+export const handleFormBlur = (e, state, setState) => {
+	const { name, value } = e.currentTarget;
+	setState({ ...state, [name]: value });
+};
